requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
tiktoken>=0.4.0; platform_system == "Windows"
faiss-cpu==1.12.0
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
sentence-transformers==2.5.1
transformers==4.38.2
datasets==2.18.0
accelerate==0.27.2
huggingface-hub==0.20.3